Extract firm types related stuff into separate class
