+++ src/main/java/minijava/ir/Collector.java
@@ -1,5 +1,7 @@
 package minijava.ir;
 
+import static minijava.ir.Types.*;
+
 import firm.*;
 import firm.Type;
 import java.util.ArrayList;
@@ -34,7 +36,7 @@ class Collector implements Program.Visitor<Void> {
       if (decl.fields.size() == 0) {
         // We have to prevent class types of length 0, so we insert an unreachable field.
         String unusableFieldName = "0padding"; // unusable because of digit prefix
-        Entity fieldEnt = new Entity(classType, unusableFieldName, IREmitter.BOOLEAN_TYPE);
+        Entity fieldEnt = new Entity(classType, unusableFieldName, BOOLEAN_TYPE);
         fieldEnt.setLdIdent(NameMangler.mangleInstanceFieldName(decl.name(), unusableFieldName));
       }
       for (Method m : decl.methods) {
@@ -50,7 +52,7 @@ class Collector implements Program.Visitor<Void> {
   }
 
   private Entity createEntity(Field f) {
-    Type type = IREmitter.getStorageType(f.type);
+    Type type = getStorageType(f.type);
     ClassType definingClass = classTypes.get(f.definingClass.def);
     Entity fieldEnt = new Entity(definingClass, f.name(), type);
     fieldEnt.setLdIdent(NameMangler.mangleInstanceFieldName(definingClass.getName(), f.name()));
@@ -72,13 +74,13 @@ class Collector implements Program.Visitor<Void> {
     parameterTypes.add(ptrTo(definingClass));
     for (LocalVariable p : m.parameters) {
       // In the body, we need to refer to local variables by index, so we save that mapping.
-      parameterTypes.add(IREmitter.getStorageType(p.type));
+      parameterTypes.add(getStorageType(p.type));
     }
 
     // The visitor returns null if that.returnType was void.
     Type[] returnTypes = {};
     if (!m.returnType.equals(minijava.ast.Type.VOID)) {
-      returnTypes = new Type[] {IREmitter.getStorageType(m.returnType)};
+      returnTypes = new Type[] {getStorageType(m.returnType)};
     }
 
     Type methodType = new MethodType(parameterTypes.toArray(new Type[0]), returnTypes);
